Major cleanup work done to remove deprecated code. Removed existing Matrix4 implementation.

+++ src/main/scala/org/sgine/core/mutable/Color.scala
@@ -3,10 +3,10 @@ package org.sgine.core.mutable
 import org.sgine.core.{Color => ImmutableColor}
 
 class Color extends ImmutableColor {
-	def red_=(red: Float) = _red = red
-	def green_=(green: Float) = _green = green
-	def blue_=(blue: Float) = _blue = blue
-	def alpha_=(alpha: Float) = _alpha = alpha
+	def red_=(red: Double) = _red = red
+	def green_=(green: Double) = _green = green
+	def blue_=(blue: Double) = _blue = blue
+	def alpha_=(alpha: Double) = _alpha = alpha
 	
 	def apply(color: Color) = {
 		red = color.red
